pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "ctf-script"
authors = [{ name = "NorthSec Challenge Designers", email = "info@nsec.io" }]
description = "NorthSec CTF challenges"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black",
    "coloredlogs==15.0.1",
    "jinja2==3.1.5",
    "jsonschema==4.23.0",
    "pyyaml<7",
    "setuptools",
    "tabulate==0.9.0",
    "typer==0.16.0",
]
version = "2.0.1"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
workflow = ["pybadges", "matplotlib", "standard-imghdr ; python_version >= \"3.13\""]

[project.scripts]
ctf = "ctf.__main__:main"

[tool.poetry]
packages = [{ include = "ctf" }]

[tool.setuptools.package-data]
ctf = ["schemas/*", "templates/*"]
